+++ package.json
@@ -39,6 +39,8 @@
     "gulp-uglify": "^1.5.3",
     "gulp-util": "^3.0.7",
     "gulp-webpack": "^1.5.0",
+    "react-redux": "^4.4.5",
+    "redux": "^3.5.2",
     "webpack": "^1.13.0",
     "webpack-dev-server": "^1.14.1"
   },
