+++ src/js/actions/index.js
@@ -2,16 +2,30 @@ export function init() {
 
 }
 
-export function center(index) {
+export function initImages(images) {
   return {
-    type: 'IMAGE_CENTER',
+    type: 'INIT_IMAGES',
+    images
+  }
+}
+
+export function centerImage(index) {
+  return {
+    type: 'CENTER_IMAGE',
     index
   };
 }
 
-export function inverse(index) {
+export function inverseImage(index) {
   return {
-    type: 'IMAGE_INVERSE',
+    type: 'INVERSE_IMAGE',
     index
+  };
+}
+
+export function rearrangeImages(images) {
+  return {
+    type: 'REARRANGE_IMAGES',
+    images
   };
 }
